requirements.txt
nltk
scikit-learn
openai
python-dotenv
spacy
flask
